Refs #21413: Move Handler-related sources to /handler
