+++ ddsrecorder/src/cpp/tool/DdsRecorder.cpp
@@ -22,12 +22,12 @@
 #include <ddspipe_core/monitoring/producers/TopicsMonitorProducer.hpp>
 #include <ddspipe_core/types/dynamic_types/types.hpp>
 
-#include <ddsrecorder_participants/recorder/mcap/McapHandler.hpp>
-#include <ddsrecorder_participants/recorder/mcap/McapHandlerConfiguration.hpp>
-#include <ddsrecorder_participants/recorder/output/BaseHandler.hpp>
+#include <ddsrecorder_participants/recorder/handler/mcap/McapHandler.hpp>
+#include <ddsrecorder_participants/recorder/handler/mcap/McapHandlerConfiguration.hpp>
+#include <ddsrecorder_participants/recorder/handler/BaseHandler.hpp>
 #include <ddsrecorder_participants/recorder/output/OutputSettings.hpp>
-#include <ddsrecorder_participants/recorder/sql/SqlHandler.hpp>
-#include <ddsrecorder_participants/recorder/sql/SqlHandlerConfiguration.hpp>
+#include <ddsrecorder_participants/recorder/handler/sql/SqlHandler.hpp>
+#include <ddsrecorder_participants/recorder/handler/sql/SqlHandlerConfiguration.hpp>
 
 #include "DdsRecorder.hpp"
