+++ ddsrecorder_participants/include/ddsrecorder_participants/recorder/handler/BaseHandlerConfiguration.hpp
