+++ ddsrecorder_participants/include/ddsrecorder_participants/recorder/handler/mcap/McapHandler.hpp
@@ -1,4 +1,4 @@
-// Copyright 2024 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+// Copyright 2023 Proyectos y Sistemas de Mantenimiento SL (eProsima).
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,19 +13,21 @@
 // limitations under the License.
 
 /**
- * @file SqlHandler.hpp
+ * @file McapHandler.hpp
  */
 
 #pragma once
 
 #include <functional>
 #include <list>
+#include <map>
 #include <memory>
-#include <set>
 #include <string>
 
+#include <mcap/mcap.hpp>
+
+#include <fastdds/dds/domain/DomainParticipantFactory.hpp>
 #include <fastdds/dds/xtypes/dynamic_types/DynamicType.hpp>
-#include <fastdds/dds/xtypes/type_representation/detail/dds_xtypes_typeobject.hpp>
 
 #include <ddspipe_core/efficiency/payload/PayloadPool.hpp>
 #include <ddspipe_core/types/data/RtpsPayloadData.hpp>
@@ -33,31 +35,30 @@
 #include <ddspipe_core/types/topic/dds/DdsTopic.hpp>
 
 #include <ddsrecorder_participants/library/library_dll.h>
+#include <ddsrecorder_participants/recorder/handler/mcap/McapHandlerConfiguration.hpp>
+#include <ddsrecorder_participants/recorder/handler/mcap/McapWriter.hpp>
 #include <ddsrecorder_participants/recorder/message/BaseMessage.hpp>
-#include <ddsrecorder_participants/recorder/message/SqlMessage.hpp>
-#include <ddsrecorder_participants/recorder/output/BaseHandler.hpp>
+#include <ddsrecorder_participants/recorder/handler/BaseHandler.hpp>
 #include <ddsrecorder_participants/recorder/output/FileTracker.hpp>
-#include <ddsrecorder_participants/recorder/sql/SqlHandlerConfiguration.hpp>
-#include <ddsrecorder_participants/recorder/sql/SqlWriter.hpp>
 
 namespace eprosima {
 namespace ddsrecorder {
 namespace participants {
 
 /**
- * Class that manages the interaction between DDS Pipe (\c SchemaParticipant) and MCAP files through sql library.
- * Payloads are efficiently passed from DDS Pipe to sql without copying data (only references).
+ * Class that manages the interaction between DDS Pipe (\c SchemaParticipant) and MCAP files through mcap library.
+ * Payloads are efficiently passed from DDS Pipe to mcap without copying data (only references).
  *
  * @implements BaseHandler
  */
-class SqlHandler : public BaseHandler
+class McapHandler : public BaseHandler
 {
 public:
 
     /**
-     * SqlHandler constructor by required values.
+     * McapHandler constructor by required values.
      *
-     * Creates SqlHandler instance with given configuration, payload pool and initial state.
+     * Creates McapHandler instance with given configuration, payload pool and initial state.
      * Opens temporal MCAP file where data is to be written.
      *
      * @throw InitializationException if creation fails (fail to open MCAP file).
@@ -67,11 +68,13 @@ public:
      *
      * @param config:       Structure encapsulating all configuration options.
      * @param payload_pool: Owner of every payload contained in received messages.
+     * @param file_tracker: File tracker to be used to create and manage MCAP files.
      * @param init_state:   Initial instance state (RUNNING/PAUSED/STOPPED).
+     * @param on_disk_full_lambda: Lambda to be executed when the disk is full.
      */
     DDSRECORDER_PARTICIPANTS_DllAPI
-    SqlHandler(
-            const SqlHandlerConfiguration& config,
+    McapHandler(
+            const McapHandlerConfiguration& config,
             const std::shared_ptr<ddspipe::core::PayloadPool>& payload_pool,
             std::shared_ptr<ddsrecorder::participants::FileTracker> file_tracker,
             const BaseHandlerStateCode& init_state = BaseHandlerStateCode::RUNNING,
@@ -85,7 +88,13 @@ public:
      *
      */
     DDSRECORDER_PARTICIPANTS_DllAPI
-    virtual ~SqlHandler();
+    virtual ~McapHandler();
+
+    /**
+     * @brief Disable MCAP handler instance
+     */
+    DDSRECORDER_PARTICIPANTS_DllAPI
+    void disable() override;
 
     /**
      * @brief Create and store in \c schemas_ an OMG IDL (.idl format) or ROS 2 (.msg format) schema.
@@ -94,6 +103,7 @@ public:
      * Previously created channels (for this type) associated with a blank schema are updated to use the new one.
      *
      * @param [in] dynamic_type DynamicType containing the type information required to generate the schema.
+     * @param [in] type_identifier  The TypeIdentifier that uniquely identifies the type in DDS systems.
      */
     DDSRECORDER_PARTICIPANTS_DllAPI
     void add_schema(
@@ -101,7 +111,7 @@ public:
             const fastdds::dds::xtypes::TypeIdentifier& type_identifier) override;
 
     /**
-     * @brief Add a data sample to the given \c topic.
+     * @brief Add a data sample, to be written through a mcap \c Channel associated to the given \c topic.
      *
      * If a channel with (non-blank) schema exists, the sample is saved in memory \c buffer_ .
      * Otherwise:
@@ -113,7 +123,7 @@ public:
      * If instance is STOPPED, received data is not processed.
      *
      * @param [in] topic DDS topic associated to this sample.
-     * @param [in] data SqlMessage to be added.
+     * @param [in] data McapMessage to be added.
      */
     DDSRECORDER_PARTICIPANTS_DllAPI
     void add_data(
@@ -125,7 +135,7 @@ protected:
     /**
      * @brief Writes \c samples to disk.
      *
-     * For each sample in \c samples, it downcasts it to \c SqlMessage, writes it to disk, and removes it from
+     * For each sample in \c samples, it downcasts it to \c McapMessage, writes it to disk, and removes it from
      * \c samples.
      * The method ends when \c samples is empty.
      *
@@ -135,28 +145,61 @@ protected:
             std::list<std::shared_ptr<const BaseMessage>>& samples) override;
 
     /**
-     * @brief Sets the key of a sample.
+     * @brief Create and add to \c mcap_writer_ channel associated to given \c topic
      *
-     * If the key was previously stored in \c keys_, it sets it.
-     * If not, it is set from the payload of the \c sql_sample.
-     * The key is stored in \c keys.
+     * A channel with blank schema is created when none found, unless only_with_schema true.
      *
-     * @param [in] sql_sample SqlMessage to set the key.
+     * @throw InconsistencyException if creation fails (schema not found and only_with_schema true).
+     *
+     * @param [in] topic Topic associated to the channel to be created
      */
-    void set_key_(
-            SqlMessage& sql_sample);
+    mcap::ChannelId create_channel_id_nts_(
+            const ddspipe::core::types::DdsTopic& topic);
+
+    /**
+     * @brief Attempt to get channel associated to given \c topic, and attempt to create one if not found.
+     *
+     * @throw InconsistencyException if not found, and creation fails (schema not found and only_with_schema true).
+     *
+     * @param [in] topic Topic associated to the channel to be created
+     */
+    mcap::ChannelId get_channel_id_nts_(
+            const ddspipe::core::types::DdsTopic& topic);
+
+    /**
+     * @brief Update channels with \c old_schema_id to use \c new_schema_id instead.
+     *
+     * Its main purpose is to update channels previously created with blank schema after having received their
+     * corresponding topic type.
+     *
+     * @param [in] old_schema_id Schema id used by the channels to be updated
+     * @param [in] new_schema_id Schema id with which to update channels (using \c old_schema_id)
+     */
+    void update_channels_nts_(
+            const mcap::SchemaId& old_schema_id,
+            const mcap::SchemaId& new_schema_id);
+
+    /**
+     * @brief Attempt to get schema with name \c schema_name.
+     *
+     * @throw InconsistencyException if not found.
+     *
+     * @param [in] schema_name Name of the schema to get.
+     */
+    mcap::SchemaId get_schema_id_nts_(
+            const std::string& schema_name);
 
     //! Configuration
-    const SqlHandlerConfiguration configuration_;
+    const McapHandlerConfiguration configuration_;
 
-    //! SQL writer
-    SqlWriter sql_writer_;
+    //! MCAP writer
+    McapWriter mcap_writer_;
 
-    //! Topics that the SQL writer has written
-    std::set<ddspipe::core::types::DdsTopic> written_topics_;
+    //! Schemas map
+    std::map<std::string, mcap::Schema> schemas_;
 
-    //! Map instance handles (hashed/serialized keys) to JSON-serialized keys
-    std::map<ddspipe::core::types::InstanceHandle, std::string> keys_;
+    //! Channels map
+    std::map<ddspipe::core::types::DdsTopic, mcap::Channel> channels_;
 };
 
 } /* namespace participants */
