+++ ddsrecorder_participants/include/ddsrecorder_participants/recorder/handler/mcap/McapSizeTracker.hpp
