+++ ddsrecorder_participants/include/ddsrecorder_participants/recorder/handler/sql/impl/SqlWriter.ipp
