+++ ddsrecorder_participants/src/cpp/recorder/handler/mcap/McapWriter.cpp
@@ -27,7 +27,7 @@
 #include <ddsrecorder_participants/recorder/exceptions/FullDiskException.hpp>
 #include <ddsrecorder_participants/recorder/exceptions/FullFileException.hpp>
 #include <ddsrecorder_participants/recorder/message/McapMessage.hpp>
-#include <ddsrecorder_participants/recorder/mcap/McapWriter.hpp>
+#include <ddsrecorder_participants/recorder/handler/mcap/McapWriter.hpp>
 
 namespace eprosima {
 namespace ddsrecorder {
