+++ ddsrecorder_participants/src/cpp/recorder/handler/sql/SqlHandler.cpp
@@ -19,8 +19,8 @@
 #define SQL_IMPLEMENTATION  // Define this in exactly one .cpp file
 
 #include <ddsrecorder_participants/common/serialize/Serializer.hpp>
+#include <ddsrecorder_participants/recorder/handler/sql/SqlHandler.hpp>
 #include <ddsrecorder_participants/recorder/message/SqlMessage.hpp>
-#include <ddsrecorder_participants/recorder/sql/SqlHandler.hpp>
 
 namespace eprosima {
 namespace ddsrecorder {
